Rename folder currency to input
